+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Changed
 -------
 - Add options to calculate variance in abundance estimates in
   the workflow ``workflow-bbduk-salmon-qc``
+- Support geneset as input to process ``mutations-table``
 
 Fixed
 -----